api/requirements.txt
chromadb
langchain
pymupdf
python-multipart
langchain-community
openai
tiktoken
fastapi
uvicorn
requests
python-dotenv